% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteColors.R
\name{gogh_interpolate}
\alias{gogh_interpolate}
\title{Interpolate the color palette}
\usage{
gogh_interpolate(palette = ggRtsy::starryNight, ...)
}
\arguments{
\item{palette}{the ggRtsy painting color palette of choice}
}
\description{
Interpolate the color palette
}
