% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteColors.R
\docType{data}
\name{gogh_palettes_pop}
\alias{gogh_palettes_pop}
\title{Different color palettes created from Van Gogh paintings}
\format{
An object of class \code{list} of length 8.
}
\usage{
gogh_palettes_pop
}
\description{
Different color palettes created from Van Gogh paintings
}
\keyword{datasets}
