% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{goghColors}
\alias{goghColors}
\title{Sampling of Colors from Van Gogh Paintings}
\format{
A data frame of 986 rows representing colors selected from
Van Gogh's painting. 5 colors were selected from each painting.
Consists of 3 variables.

\describe{
  \item{name}{Color name}
  \item{rgb}{Color RGB value}
  \item{hex}{Color hex code value}
}
}
\usage{
goghColors
}
\description{
Sampling of Colors from Van Gogh Paintings
}
\keyword{datasets}
