% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{goghPaintingSets}
\alias{goghPaintingSets}
\title{Van Gogh Paintings Information}
\format{
A data frame of 1931 rows representing information from
Van Gogh's paintings. Consists of 6 variables.

\describe{
  \item{name}{Painting name}
  \item{colors}{5 colors picked from the painting}
  \item{year}{Year painting was painted}
  \item{genre}{Genre of painting}
  \item{style}{Style of painting}
  \item{link}{Link of image of painting}
}
}
\source{
Kaggle Konstantinos Katserelis <https://www.kaggle.com/datasets/pointblanc/colors-of-van-gogh?select=color_space.csv>
Note data is released under CC BY-SA 4.0
}
\usage{
goghPaintingSets
}
\description{
This dataset consists of information about and colors from all 1931
of Van Gogh's paintings.
}
\keyword{datasets}
