% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGBToHex.R
\name{rgbToHex}
\alias{rgbToHex}
\title{Converts RGB colors to Hex code colors}
\usage{
rgbToHex(rgbVec)
}
\arguments{
\item{rgbVec}{a vector of strings in the format "(0, 72, 186)"}
}
\value{
vector of hex code strings in the format "#0048ba"
}
\description{
Converts RGB colors to Hex code colors
}
\details{
Taking input of a vector of RGB values in the format
"(0, 72, 186)", etc, and returns the corresponding
hex value in the format "#0048ba"
}
