% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{goghColors}
\alias{goghColors}
\title{Sampling of Colors from Van Gogh Paintings}
\format{
A data frame of 986 rows representing colors selected from
Van Gogh's painting. 5 colors were selected from each painting.
Consists of 3 variables.

\describe{
  \item{name}{Color name}
  \item{rgb}{Color RGB value}
  \item{hex}{Color hex code value}
}
}
\source{
Kaggle Konstantinos Katserelis <https://www.kaggle.com/datasets/pointblanc/colors-of-van-gogh?select=color_space.csv>
Note data is released under CC BY-SA 4.0
}
\usage{
goghColors
}
\description{
This dataset consists of information on 986 colors across Van Gogh paintings.
}
\keyword{datasets}
